pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matphase"
version = "0.1.0"
description = "Python toolkit for fMRI phase-field spiral detection and analysis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "MatPhase Contributors"}
]
keywords = ["fMRI", "neuroimaging", "phase-field", "spiral-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "numpy>=2.0",
    "scipy>=1.13",
    "nibabel>=3.2",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "h5py>=3.12",
]

[project.optional-dependencies]
gpu = [
    "cupy>=10.0",
    "taichi>=1.6",
]
viz = [
    "matplotlib>=3.8",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "ruff>=0.1",
    "black>=23.0",
    "mypy>=1.0",
]
all = [
    "matphase[gpu,viz,dev]",
]

[project.scripts]
matphase = "matphase.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/matphase"
Repository = "https://github.com/your-org/matphase"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
python_classes = ["Test*"]
python_functions = ["test_*"]
